PythonSerialProxy/requirements.txt
firebase_admin
pyserial
google.generativeai
python-dotenv